requirements.txt
flask>=3.0.0
requests>=2.31.0
pyyaml>=6.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
pytest-cov>=4.1.0
responses>=0.25.0
pytest-playwright>=0.4.0
playwright>=1.40.0
pypdf>=4.0.0
